Gem Fighter, not Puzzle Fighter

+++ palmod/Game/Game_COTA_A.cpp
@@ -773,7 +773,7 @@ void CGame_COTA_A::UpdatePalData()
             {
                 UINT16 nCurrentColorCountToWrite = min(nMaxSafeColorsToWrite, nTotalColorsRemaining);
 
-                for (int nPICtr = 0; nPICtr < nCurrentColorCountToWrite; nPICtr++)
+                for (UINT16 nPICtr = 0; nPICtr < nCurrentColorCountToWrite; nPICtr++)
                 {
                     if (nPICtr == iFixedCounterPosition)
                     {
