+++ palmod/Game/Game_KOF02UM_S.cpp
@@ -913,7 +913,7 @@ void CGame_KOF02UM_S::UpdatePalData()
             {
                 UINT16 nCurrentColorCountToWrite = min(nMaxSafeColorsToWrite, nTotalColorsRemaining);
 
-                for (int nPICtr = 0; nPICtr < nCurrentColorCountToWrite; nPICtr++)
+                for (UINT16 nPICtr = 0; nPICtr < nCurrentColorCountToWrite; nPICtr++)
                 {
                     if (nPICtr == iFixedCounterPosition)
                     {
