+++ palmod/Game/Game_MSHVSF_A.cpp
@@ -925,7 +925,7 @@ void CGame_MSHVSF_A::UpdatePalData()
             {
                 UINT16 nCurrentColorCountToWrite = min(nMaxSafeColorsToWrite, nTotalColorsRemaining);
 
-                for (int nPICtr = 0; nPICtr < nCurrentColorCountToWrite; nPICtr++)
+                for (UINT16 nPICtr = 0; nPICtr < nCurrentColorCountToWrite; nPICtr++)
                 {
                     if (nPICtr == iFixedCounterPosition)
                     {
