+++ palmod/Game/Game_SFIII3_D.cpp
@@ -328,7 +328,7 @@ void CGame_SFIII3_D::UpdatePalData()
             UINT16 uAmt = srcDef->uPalSz;
 
             // First color is the transparency color
-            for (int nPICtr = 1; nPICtr < uAmt; nPICtr++)
+            for (UINT16 nPICtr = 1; nPICtr < uAmt; nPICtr++)
             {
                 pppDataBuffer[srcDef->uUnitId][srcDef->uPalId][nPICtr] = (ConvCol(crSrc[nPICtr]) | 0x8000);
             }
